Sell to Stable Coin (of choice) feature
I am looking for the ability to remove certain aspects of market risk.  These coins are all coupled together (currently). Panic selling into BTC reduces alt coin exposure but not nearly to the degree a stable coin would. 

There should be a one click button or a stop loss on BTC (%) that triggers the bot to sell all coins into a stable coin of choice (TrueUSD).  Depending on the approach, another button (business as usual) or a BTC recover % that would function in the same way as the button.